+++ lib/overcommit/hook/pre_commit/author_email.rb
@@ -6,9 +6,10 @@ module Overcommit::Hook::PreCommit
       email = result.stdout.chomp
 
       unless email =~ /#{@config['pattern']}/
-        return :bad, "Author has an invalid email address: '#{email}'\n" \
-                     'Set your email with ' \
-                     '`git config --global user.email your_email@example.com`'
+        return :fail,
+               "Author has an invalid email address: '#{email}'\n" \
+               'Set your email with ' \
+               '`git config --global user.email your_email@example.com`'
       end
 
       :good