+++ lib/overcommit/hook/pre_commit/author_name.rb
@@ -6,9 +6,9 @@ module Overcommit::Hook::PreCommit
       name = result.stdout.chomp
 
       unless name.split(' ').count >= 2
-        return :bad, 'Author must have at least first and last name, but ' \
-                     "was: '#{name}'.\nSet your name with " \
-                     "`git config --global user.name 'Your Name'`"
+        return :fail,
+               "Author must have at least first and last name, but was: #{name}." \
+               "\nSet your name with `git config --global user.name 'Your Name'`"
       end
 
       :good