+++ lib/overcommit/hook/pre_commit/chamber_security.rb
@@ -9,7 +9,7 @@ module Overcommit::Hook::PreCommit
       result = execute(%w[chamber secure --echo --files] + applicable_files)
 
       return :good if result.stdout.empty?
-      [:bad, "These settings appear to need to be secured but were not: #{result.stdout}"]
+      [:fail, "These settings appear to need to be secured but were not: #{result.stdout}"]
     end
   end
 end