+++ lib/overcommit/hook/pre_commit/json_syntax.rb
@@ -16,7 +16,7 @@ module Overcommit::Hook::PreCommit
 
       return :good if output.empty?
 
-      [:bad, output]
+      [:fail, output]
     end
   end
 end