+++ lib/overcommit/hook/pre_commit/travis_lint.rb
@@ -9,7 +9,7 @@ module Overcommit::Hook::PreCommit
       result = execute(%w[travis-lint] + applicable_files)
       return :good if result.success?
 
-      [:bad, result.stdout.strip]
+      [:fail, result.stdout.strip]
     end
   end
 end