+++ spec/support/matchers/hook.rb
@@ -40,7 +40,7 @@ end
 
 # Can't use 'fail' as it is a reserved word.
 RSpec::Matchers.define :fail_hook  do |*args|
-  check_matcher = HookMatcher.new(:bad, args)
+  check_matcher = HookMatcher.new(:fail, args)
 
   match do
     check_matcher.matches?(actual)